+++ src/server/domain/item.cpp
@@ -5,8 +5,9 @@
 namespace asciinem::server::domain
 {
 
-item::item( int id, std::string name, int value, int level )
-    : id_( id ), name_( std::move( name ) ), value_( value ), level_( level )
+item::item( int id, std::string name, double value, int level )
+    : id_( id ), name_( std::move( name ) ), value_( money( value ) ),
+      level_( level )
 {
 }
 
@@ -15,7 +16,7 @@ auto item::get_name() const -> const std::string&
     return name_;
 }
 
-auto item::get_value() const -> int
+auto item::get_value() const -> money
 {
     return value_;
 }
@@ -25,9 +26,9 @@ auto item::get_level() const -> int
     return level_;
 }
 
-void item::set_value( int value )
+void item::set_value( double value )
 {
-    value_ = value;
+    value_ = money( value );
 }
 
 void item::set_name( const std::string& name )
